Fix SearchRequest scope property
The `.scope` property should not return a name. It should return the
actual code so that it can be compared against the RFC defined
integers for scopes. We replicated the original implementation without
understanding how it was used. This was incorrect.

We also add a `.scopeName` property so that the name can be retrieved
if desired.